+++ src/app/actions/badge-notification.ts
@@ -1,12 +1,12 @@
 "use server";
 
 import { markBadgeNotificationAsSeen } from "@/lib/services/badgeNotification";
-import { createClient } from "@/lib/supabase/server";
+import { createClient } from "@/lib/supabase/client";
 
 export async function markBadgeNotificationAsSeenAction(
   badgeIds: string[],
 ): Promise<{ success: boolean; error?: string }> {
-  const supabase = await createClient();
+  const supabase = createClient();
   const {
     data: { user },
   } = await supabase.auth.getUser();
