+++ src/app/map/posting/page.tsx
@@ -1,4 +1,4 @@
-import { createClient } from "@/lib/supabase/server";
+import { createClient } from "@/lib/supabase/client";
 import type { Metadata } from "next";
 import { redirect } from "next/navigation";
 import PostingPageClient from "./PostingPageClient";
@@ -9,7 +9,7 @@ export const metadata: Metadata = {
 };
 
 export default async function PostingPage() {
-  const supabase = await createClient();
+  const supabase = createClient();
 
   const {
     data: { user },
