+++ src/components/navbar.tsx
@@ -8,13 +8,13 @@ import {
   DropdownMenuSeparator,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
-import { createClient } from "@/lib/supabase/server";
+import { createClient } from "@/lib/supabase/client";
 import { Menu } from "lucide-react";
 import Image from "next/image";
 import Link from "next/link";
 
 export default async function Navbar() {
-  const supabase = await createClient();
+  const supabase = createClient();
 
   const {
     data: { user },
