+++ src/lib/services/missionsRanking.test.ts
@@ -1,10 +1,10 @@
-import { createClient } from "@/lib/supabase/server";
+import { createClient } from "@/lib/supabase/client";
 import { getJSTMidnightToday } from "../dateUtils";
 import { getMissionRanking, getUserMissionRanking } from "./missionsRanking";
 import { getCurrentSeasonId } from "./seasons";
 
 // Supabaseクライアントをモック
-jest.mock("@/lib/supabase/server", () => ({
+jest.mock("@/lib/supabase/client", () => ({
   createClient: jest.fn(),
 }));
 
@@ -20,7 +20,7 @@ describe("missionsRanking service", () => {
 
   beforeEach(() => {
     jest.clearAllMocks();
-    (createClient as jest.Mock).mockResolvedValue(mockSupabase);
+    (createClient as jest.Mock).mockReturnValue(mockSupabase);
     (getCurrentSeasonId as jest.Mock).mockResolvedValue("test-season-id");
   });
