+++ src/lib/services/seasons.ts
@@ -1,12 +1,12 @@
 import { createAdminClient } from "@/lib/supabase/adminClient";
-import { createClient } from "@/lib/supabase/server";
+import { createClient } from "@/lib/supabase/client";
 import type { Database } from "@/lib/types/supabase";
 
 type Season = Database["public"]["Tables"]["seasons"]["Row"];
 export type { Season };
 
 export async function getCurrentSeason(): Promise<Season | null> {
-  const supabase = await createClient();
+  const supabase = createClient();
 
   const { data, error } = await supabase
     .from("seasons")
@@ -23,7 +23,7 @@ export async function getCurrentSeason(): Promise<Season | null> {
 }
 
 export async function getAllSeasons(): Promise<Season[]> {
-  const supabase = await createClient();
+  const supabase = createClient();
 
   const { data, error } = await supabase
     .from("seasons")
@@ -40,7 +40,7 @@ export async function getAllSeasons(): Promise<Season[]> {
 }
 
 export async function getInactiveSeasons(): Promise<Season[]> {
-  const supabase = await createClient();
+  const supabase = createClient();
 
   const { data, error } = await supabase
     .from("seasons")
@@ -57,7 +57,7 @@ export async function getInactiveSeasons(): Promise<Season[]> {
 }
 
 export async function getSeasonBySlug(slug: string): Promise<Season | null> {
-  const supabase = await createClient();
+  const supabase = createClient();
 
   const { data, error } = await supabase
     .from("seasons")
