--- src/lib/supabase/server.ts
@@ -1,42 +1,0 @@
-"use server";
-
-import type { Database } from "@/lib/types/supabase";
-import { createServerClient } from "@supabase/ssr";
-
-import { cookies } from "next/headers";
-
-export const createClient = async () => {
-  const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL;
-  const supabaseAnonKey = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY;
-
-  if (!supabaseUrl) {
-    throw new Error("Missing NEXT_PUBLIC_SUPABASE_URL environment variable");
-  }
-
-  if (!supabaseAnonKey) {
-    throw new Error(
-      "Missing NEXT_PUBLIC_SUPABASE_ANON_KEY environment variable",
-    );
-  }
-
-  return createServerClient<Database>(supabaseUrl, supabaseAnonKey, {
-    cookies: {
-      async getAll() {
-        const cookieStore = await cookies();
-        return cookieStore.getAll();
-      },
-      async setAll(cookiesToSet) {
-        try {
-          const cookieStore = await cookies();
-          for (const { name, value, options } of cookiesToSet) {
-            cookieStore.set(name, value, options);
-          }
-        } catch (error) {
-          // The `set` method was called from a Server Component.
-          // This can be ignored if you have middleware refreshing
-          // user sessions.
-        }
-      },
-    },
-  });
-};
